Implement support for <content:encoded> (fixes #4)
Implement a new property FPItem.description. This takes the place of the old
FPItem.content. FPItem.content is now used for <content:encoded>. If the item
does not contain a <content:encoded> tag, this property will contain the value
of item.description. This is a mildly backwards-incompatible change, in that
if a client of FeedParser was using FPItem.content along with the extension
element mechanism to extract both <description> and <content:encoded>, that
code will no longer work correctly.

The explicit handling of <content:encoded> removes it from the extension
element mechanism. A solution to this is forthcoming. In the meantime,
some tests are expected to fail.