Support odd date format used by Google News (fixes #8)
Google News has a weird quirk with its RSS dates where it gives
every date both a textual timezone as well as a numeric one.
In addition, it adds a separator between the hours and minutes
in the numeric timezone. Neither of these quirks are RFC822-compliant.
We should support them anyway.