Change our class prefix to FDP
The FP prefix is used internally by FileProvider.framework. In
particular, it defines an `FPItem` that conflicts with ours.

Use `@compatibility_alias` for all of the classes, and provide
deprecated renames for almost everything else. The only thing we can't
provide a compatibility path for is the `FDPXMLParserProtocol` protocol,
but that's an internal implementation detail so it doesn't matter.